Struct open_graph::Html

pub struct Html {
    pub errors: Vec<Cow<'static, str>, Global>,
    pub quirks_mode: QuirksMode,
    pub tree: Tree<Node>,
}

An HTML tree.

Parsing does not fail hard. Instead, the quirks_mode is set and errors are added to the errors field. The tree will still be populated as best as possible.

Implements the TreeSink trait from the html5ever crate, which allows HTML to be parsed.

Fields
errors: Vec<Cow<'static, str>, Global>

Parse errors.

quirks_mode: QuirksMode

The quirks mode.

tree: Tree<Node>

The node tree.

Implementations

impl Html

pub fn new_document() -> Html

Creates an empty HTML document.

pub fn new_fragment() -> Html

Creates an empty HTML fragment.

pub fn parse_document(document: &str) -> Html

Parses a string of HTML as a document.

This is a convenience method for the following:

use html5ever::driver::{self, ParseOpts};
use scraper::Html;
use tendril::TendrilSink;

let parser = driver::parse_document(Html::new_document(), ParseOpts::default());
let html = parser.one(document);

pub fn parse_fragment(fragment: &str) -> Html

Parses a string of HTML as a fragment.

pub fn select(&'a self, selector: &'b Selector) -> Select<'a, 'b>

Returns an iterator over elements matching a selector.

pub fn root_element(&self) -> ElementRef<'_>

Returns the root <html> element.